# docs/api/public-api.openapi.yaml
openapi: 3.0.3
info:
  title: OpenGlass Public API
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
  description: |
    REST contract for the OpenGlass public layer (MVP scope).

    Source of truth for both `MockApiClient` and the Go backend — see
    `docs/dev/frontend-first-development.md`. No invented wire fields:
    additions require amending this file first.

    Conventions:
    - All endpoints are under `/api/v1`.
    - Invite issuance is an **ops-side operation** (server CLI / admin
      tooling on the instance host), deliberately not part of the public
      client API — there is no "create invite" endpoint here.
    - Auth uses `Authorization: Bearer <access_token>` (JWT, short-lived)
      plus a refresh token cookie/body flow.
    - Error envelope and codes: `docs/api/errors.md`.
    - Realtime events are out of scope here — see `docs/api/ws-events.md`.
servers:
  - url: /api/v1
...
paths:
...
  /auth/otp/request:
    post:
      tags: [auth]
      summary: Request an OTP (S1 → S2)
      security: []
      operationId: requestOtp
      description: |
        Validates the invite code and sends a one-time code to the email.
        Rate-limited; resend cooldown is a contract constant
        (`OTP_RESEND_COOLDOWN_S = 60`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                inviteCode:
                  type: string
                  example: "GLS-9F2K-71"
                  description: |
                    Required for first registration only. Returning users
                    re-login with email + OTP alone — the invite gates
                    account creation, not every sign-in. Omitting it for an
                    unknown email → `invite_required`. The invite is
                    consumed at successful `verifyOtp`, not at this step —
                    a failed OTP attempt does not burn it.
                email: { type: string, format: email }
      responses:
        '202':
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                required: [otpExpiresInS, resendAvailableInS]
                properties:
                  otpExpiresInS: { type: integer, example: 600 }
                  resendAvailableInS: { type: integer, example: 60 }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '429':
          description: Resend cooldown active (`otp_cooldown`)
          headers:
            Retry-After:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: otp_cooldown
                  message: "Wait before requesting a new code"
                  details: { resendAvailableInS: 42 }
        '403':
          description: |
            `invite_required` — email unknown and no invite supplied;
            `invite_invalid` — invite supplied but wrong/consumed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                required:
                  value: { error: { code: invite_required, message: "Invite code required to register" } }
                invalid:
                  value: { error: { code: invite_invalid, message: "Invite code is invalid or already used" } }
  /auth/otp/verify:
    post:
      tags: [auth]
      summary: Verify OTP (S2 → S3 or straight to S4 for existing users)
      security: []
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, deviceName]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: '^[0-9]{6}$' }
                deviceName: { type: string, example: "Pixel 8 — Chrome PWA" }
      responses:
        '200':
          description: |
            Verified. `needsProfile=true` for first login (client shows S3);
            `false` for returning users (client goes to S4).
          content:
            application/json:
              schema:
                type: object
                required: [needsProfile, accessToken, accessTokenExpiresInS, refreshToken]
                properties:
                  needsProfile: { type: boolean }
                  accessToken: { type: string }
                  accessTokenExpiresInS: { type: integer, example: 900 }
                  refreshToken: { type: string }
                  user:
                    type: object
                    nullable: true
                    description: |
                      `User` object, present only when `needsProfile=false`.
                      First login has no displayName/tag yet — profile is
                      created via `POST /auth/profile` (S3).
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403':
          description: Code invalid or expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                invalid:
                  value: { error: { code: otp_invalid, message: "Wrong code" } }
                expired:
                  value: { error: { code: otp_expired, message: "Code expired, request a new one" } }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/profile:
    post:
      tags: [auth]
      summary: Complete profile setup (S3 → S4)
      operationId: completeProfile
      description: |
        Sets the display name and binds the immutable unique tag
        `name#NNNN`. Requires `needsProfile` session. If the requested tag
        prefix is taken, `tagSuggestions` are returned instead. One-shot:
        on an account that already has a profile this returns
        `conflict` — use `PATCH /users/me` for later display-name changes
        (the tag never changes). When `requestedTag` is omitted the
        display-name prefix is attempted; a collision also yields
        `tag_taken` + suggestions.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName]
              properties:
                displayName: { type: string, minLength: 1, maxLength: 64, example: "jojou" }
                requestedTag:
                  type: string
                  description: Optional explicit tag attempt (e.g. "Jojou13245" from suggestions)
      responses:
        '200':
          description: Profile created
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '409':
          description: Tag taken — suggestions in `error.details.tagSuggestions`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                tag_taken:
                  value:
                    error:
                      code: tag_taken
                      message: "Tag is taken, pick a suggestion"
                      details:
                        tagSuggestions: ["Jojou13245", "Jojou48127", "jojou#9031"]
                conflict:
                  value:
                    error:
                      code: conflict
                      message: "Profile already exists"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate access + refresh tokens
      description: |
        Refresh tokens rotate on every use — the presented token is
        invalidated and a new pair is issued. Presenting an already-used
        refresh token (reuse detected, possible theft) revokes the entire
        device session: client gets `unauthorized` and must log in again.
      security: []
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, accessTokenExpiresInS, refreshToken]
                properties:
                  accessToken: { type: string }
                  accessTokenExpiresInS: { type: integer }
                  refreshToken: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke the current device session
      description: |
        Revokes this device's session; Web Push subscriptions bound to it
        are removed with it.
      operationId: logout
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Session revoked }

        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/sessions:
    get:
      tags: [auth]
      summary: List device sessions (S14)
      operationId: listSessions
      security: [{ bearerAuth: [] }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: { $ref: '#/components/schemas/DeviceSession' }
  /auth/sessions/{sessionId}:
    delete:
      tags: [auth]
      summary: Revoke a specific device session (S14)
      operationId: revokeSession
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/SessionId' }
      responses:
        '204': { description: Revoked }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/me:
    get:
      tags: [users]
      summary: Own profile (app cold start, S4)
      operationId: getMe
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current user + private fields
          content:
            application/json:
              schema:
                type: object
                required: [user, email]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  email:
                    type: string
                    format: email
                    description: Own login email — private, never exposed
                      on the public `User` object or to other users
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [users]
      summary: Update own profile (S13 — display name is changeable anytime)
      operationId: updateMe
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string, minLength: 1, maxLength: 64 }
                avatarUrl: { type: string, nullable: true }
      responses:
        '200':
          description: Updated profile. `tag` never changes.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/search:
    get:
      tags: [users]
      summary: Search users by display name or tag (S6)
      operationId: searchUsers
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
          description: Display-name substring or exact `name#NNNN` tag
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
  /users/{userId}:
    get:
      tags: [users]
      summary: User profile card (S7)
      operationId: getUser
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/UserId' }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [user, relationship]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  verifiedAt:
                    type: string
                    format: date-time
                    nullable: true
                    description: |
                      Last identity verification of this contact (US-2.5
                      "verified" mark on the profile card). Always null
                      until the private module ships.
                  relationship:
                    type: string
                    enum: [none, contact_incoming, contact_outgoing, contact_mutual, self]
                    description: |
                      none — strangers; contact_incoming — they added you,
                      you haven't (drives "add back?" prompt); contact_outgoing —
                      you added them; contact_mutual — both added; self.
        '404': { $ref: '#/components/responses/NotFound' }
  /contacts:
    get:
      tags: [contacts]
      summary: Contact list
      operationId: listContacts
      security: [{ bearerAuth: [] }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [contacts]
                properties:
                  contacts:
                    type: array
                    items: { $ref: '#/components/schemas/Contact' }
    post:
      tags: [contacts]
      summary: Add contact (S7)
      operationId: addContact
      description: |
        Creates a one-directional contact entry. When both users have added
        each other the relation becomes `mutual` — required for 1-1 chat
        creation and private-session invites.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [contact]
                properties:
                  contact: { $ref: '#/components/schemas/Contact' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already in contacts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /contacts/{userId}:
    delete:
      tags: [contacts]
      summary: Remove contact
      operationId: removeContact
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/UserId' }
      responses:
        '204': { description: Removed }
        '404': { $ref: '#/components/responses/NotFound' }
  /chats:
    get:
      tags: [chats]
      summary: Chat list (S4)
      operationId: listChats
      description: Ordered by last activity; includes 1-1 chats and groups.
      security: [{ bearerAuth: [] }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [chats]
                properties:
                  chats:
                    type: array
                    items: { $ref: '#/components/schemas/ChatSummary' }
    post:
      tags: [chats]
      summary: Open or create a 1-1 chat with a mutual contact (S7 → S5)
      operationId: openDirectChat
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '200':
          description: Existing chat returned
          content:
            application/json:
              schema:
                type: object
                required: [chat]
                properties:
                  chat: { $ref: '#/components/schemas/Chat' }
        '201':
          description: New chat created
          content:
            application/json:
              schema:
                type: object
                required: [chat]
                properties:
                  chat: { $ref: '#/components/schemas/Chat' }
        '403':
          description: Not mutual contacts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /chats/{chatId}:
    get:
      tags: [chats]
      summary: Chat details (S5 header / S7 actions)
      operationId: getChat
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [chat]
                properties:
                  chat: { $ref: '#/components/schemas/Chat' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chats/{chatId}/pin:
    post:
      tags: [chats]
      summary: Pin/unpin chat in own chat list (S4 ordering)
      operationId: setChatPinned
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pinned]
              properties:
                pinned: { type: boolean }
      responses:
        '204':
          description: Updated
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chats/{chatId}/messages:
    get:
      tags: [messages]
      summary: Message history, cursor-paginated (S5)
      operationId: listMessages
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
        - name: before
          in: query
          schema: { type: string }
          description: Opaque cursor from `nextCursor` of a previous page
            (older messages). Server implements it over the per-chat `seq`;
            clients must treat it as opaque.
        - name: after
          in: query
          schema: { type: string }
          description: Opaque cursor from `newerCursor` — paginates toward
            newer messages (forward scroll after an `around` jump).
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
        - name: q
          in: query
          schema: { type: string, minLength: 1 }
          description: Full-text search within this chat (S5b). When set,
            results are matching messages ordered by seq desc; `before`
            cursor paginates further matches.
        - name: pinned
          in: query
          schema: { type: boolean }
          description: When true, returns only pinned messages (S5b pinned view)
        - name: around
          in: query
          schema: { type: string, format: uuid }
          description: |
            Jump-to-message (search results, pinned messages, replies):
            returns a window of `limit` messages centered on this message's
            seq. Mutually exclusive with `q` and `pinned`.
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [messages, nextCursor, newerCursor]
                properties:
                  messages:
                    type: array
                    items: { $ref: '#/components/schemas/Message' }
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor toward older messages; null at history start
                  newerCursor:
                    type: string
                    nullable: true
                    description: |
                      Cursor toward newer messages — only non-null after an
                      `around` jump, letting the client scroll forward from
                      the jumped position; pass it in the `after` param.
                      null when already at the latest.
    post:
      tags: [messages]
      summary: Send a message (S5)
      operationId: sendMessage
      description: |
        Server assigns `id`, `seq`, `sentAt`. Delivery to online members
        happens via `message.new` WS event — see `docs/api/ws-events.md`.
        Optimistic-UI/outbox semantics are defined by the contract: client
        sends `clientNonce`; server echoes it in the response and the WS
        event for dedup. The pair `(senderId, clientNonce)` is unique:
        retrying a send with the same nonce returns the already-created
        message with `200` (never a duplicate), so outbox flush is safe.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientNonce]
              properties:
                clientNonce:
                  type: string
                  description: Client-generated idempotency key (uuid)
                text:
                  type: string
                  maxLength: 8192
                  description: Required unless `attachmentIds` is non-empty
                attachmentIds:
                  type: array
                  items: { type: string, format: uuid }
                  maxItems: 10
                  description: |
                    Required unless `text` is non-empty. Each id must
                    reference an attachment uploaded to this chat
                    (`POST /chats/{chatId}/attachments`), else
                    `validation_failed`.
                replyToMessageId:
                  type: string
                  format: uuid
                  description: Must reference a message in this chat
      responses:
        '200':
          description: Idempotent replay — `clientNonce` was already used,
            returns the existing message
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
  /messages/{messageId}:
    patch:
      tags: [messages]
      summary: Edit own message
      operationId: editMessage
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/MessageId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1, maxLength: 8192 }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [messages]
      summary: Delete a message
      operationId: deleteMessage
      description: Own messages; in groups also users with `delete_messages` right.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/MessageId' }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /messages/{messageId}/pin:
    post:
      tags: [messages]
      summary: Pin/unpin a message (S5b pinned view)
      description: |
        Direct chats: either party. Groups: requires `pin_messages`
        right or owner.
      operationId: setMessagePinned
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/MessageId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pinned]
              properties:
                pinned: { type: boolean }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chats/{chatId}/attachments:
    post:
      tags: [messages]
      summary: Upload an attachment (photo/file)
      operationId: uploadAttachment
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [attachment]
                properties:
                  attachment: { $ref: '#/components/schemas/Attachment' }
        '413':
          description: File too large
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /chats/{chatId}/read:
    post:
      tags: [messages]
      summary: Mark chat read up to a message (read receipts)
      operationId: markRead
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upToSeq]
              properties:
                upToSeq: { type: integer }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '204': { description: Recorded }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationFailed:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit hit
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: rate_limited, message: "Too many requests" } }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
...